Other classes adapted.
Now, the program can change its way to work between the two graphs types by changing a variable in the code.

Also, all the game is playable by using the two graphs types.

It just remains to implement the asked graphs algorithms in the game and it leaves for painting (sale pa pintura xd).